Refactoring the files to put TreeListView in its own resource dictionary
